+++ sertop/sertop_util.mli
@@ -31,3 +31,5 @@ type fb_filter_opts = {
 val default_fb_filter_opts : fb_filter_opts
 
 val feedback_opt_filter : ?opts:fb_filter_opts -> Feedback.feedback -> Feedback.feedback option
+
+val feedback_tr : Feedback.feedback -> Serapi_protocol.feedback